# contract/openapi.json
{
  "components" : {
    "schemas" : {
...
      "BudgetRequest" : {
        "properties" : {
          "amountLimit" : {
            "minimum" : 0.0,
            "type" : "number"
          },
          "categoryId" : {
            "format" : "int64",
            "type" : "integer"
          },
          "currency" : {
            "type" : "string"
          },
          "exchangeRate" : {
            "minimum" : 0.0,
            "type" : "number"
          },
          "month" : {
            "minLength" : 1,
            "pattern" : "\\d{4}-(0[1-9]|1[0-2])",
            "type" : "string"
          },
          "recurring" : {
            "type" : "boolean"
          }
        },
        "required" : [ "amountLimit", "categoryId", "month" ],
        "type" : "object"
      },
...
      "BudgetResponse" : {
        "properties" : {
          "amountLimit" : {
            "type" : "number"
          },
          "amountLimitInBaseCurrency" : {
            "type" : "number"
          },
          "categoryId" : {
            "format" : "int64",
            "type" : "integer"
          },
          "categoryName" : {
            "type" : "string"
          },
          "currency" : {
            "type" : "string"
          },
          "exchangeRate" : {
            "type" : "number"
          },
          "id" : {
            "format" : "int64",
            "type" : "integer"
          },
          "month" : {
            "type" : "string"
          },
          "recurring" : {
            "type" : "boolean"
          }
        },
        "type" : "object"
      },
...
      "CategoryRequest" : {
        "properties" : {
          "icon" : {
            "maxLength" : 50,
            "minLength" : 0,
            "type" : "string"
          },
          "name" : {
            "maxLength" : 50,
            "minLength" : 0,
            "type" : "string"
          }
        },
        "required" : [ "name" ],
        "type" : "object"
      },
      "CategoryResponse" : {
        "properties" : {
          "bucket" : {
            "enum" : [ "NEEDS", "WANTS", "SAVINGS" ],
            "type" : "string"
          },
          "icon" : {
            "type" : "string"
          },
          "id" : {
            "format" : "int64",
            "type" : "integer"
          },
          "name" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "ChatMessageDto" : {
        "properties" : {
          "content" : {
            "type" : "string"
          },
          "createdAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "id" : {
            "format" : "int64",
            "type" : "integer"
          },
          "responseType" : {
            "type" : "string"
          },
          "role" : {
            "type" : "string"
          },
          "toolName" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "ConversationSummaryDto" : {
        "properties" : {
          "createdAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "id" : {
            "format" : "int64",
            "type" : "integer"
          },
          "title" : {
            "type" : "string"
          },
          "updatedAt" : {
            "format" : "date-time",
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "ExpenseRequest" : {
        "properties" : {
          "amount" : {
            "minimum" : 0.0,
            "type" : "number"
          },
          "category" : {
            "maxLength" : 50,
            "minLength" : 0,
            "type" : "string"
          },
          "currency" : {
            "maxLength" : 3,
            "minLength" : 0,
            "type" : "string"
          },
          "date" : {
            "format" : "date-time",
            "type" : "string"
          },
          "description" : {
            "minLength" : 1,
            "type" : "string"
          },
          "exchangeRate" : {
            "minimum" : 1.0E-6,
            "type" : "number"
          },
          "expenseType" : {
            "type" : "string"
          },
          "reimbursable" : {
            "type" : "boolean"
          }
        },
        "required" : [ "amount", "description" ],
        "type" : "object"
      },
...
      "ExpenseResponse" : {
        "properties" : {
          "amount" : {
            "type" : "number"
          },
          "amountInBaseCurrency" : {
            "type" : "number"
          },
          "category" : {
            "type" : "string"
          },
          "currency" : {
            "type" : "string"
          },
          "date" : {
            "format" : "date-time",
            "type" : "string"
          },
          "description" : {
            "type" : "string"
          },
          "exchangeRate" : {
            "type" : "number"
          },
          "expenseType" : {
            "type" : "string"
          },
          "id" : {
            "format" : "int64",
            "type" : "integer"
          },
          "reimbursable" : {
            "type" : "boolean"
          }
        },
        "type" : "object"
      },
...
      "RecurringExpenseRequest" : {
        "properties" : {
          "active" : {
            "type" : "boolean"
          },
          "amount" : {
            "minimum" : 0.0,
            "type" : "number"
          },
          "categoryName" : {
            "type" : "string"
          },
          "currency" : {
            "type" : "string"
          },
          "dayOfMonth" : {
            "format" : "int32",
            "type" : "integer"
          },
          "dayOfWeek" : {
            "format" : "int32",
            "type" : "integer"
          },
          "exchangeRate" : {
            "type" : "number"
          },
          "frequency" : {
            "enum" : [ "MONTHLY", "WEEKLY", "YEARLY" ],
            "type" : "string"
          },
          "monthOfYear" : {
            "format" : "int32",
            "type" : "integer"
          },
          "name" : {
            "maxLength" : 100,
            "minLength" : 0,
            "type" : "string"
          }
        },
        "required" : [ "amount", "frequency", "name" ],
        "type" : "object"
      },
...
      "RecurringExpenseResponse" : {
        "properties" : {
          "active" : {
            "type" : "boolean"
          },
          "amount" : {
            "type" : "number"
          },
          "categoryName" : {
            "type" : "string"
          },
          "currency" : {
            "type" : "string"
          },
          "dayOfMonth" : {
            "format" : "int32",
            "type" : "integer"
          },
          "dayOfWeek" : {
            "format" : "int32",
            "type" : "integer"
          },
          "exchangeRate" : {
            "type" : "number"
          },
          "frequency" : {
            "enum" : [ "MONTHLY", "WEEKLY", "YEARLY" ],
            "type" : "string"
          },
          "id" : {
            "format" : "int64",
            "type" : "integer"
          },
          "monthOfYear" : {
            "format" : "int32",
            "type" : "integer"
          },
          "name" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
    },
    "securitySchemes" : {
      "bearerAuth" : {
        "bearerFormat" : "JWT",
        "description" : "Backend-issued JWT from POST /auth/login, sent as `Authorization: Bearer <token>`.",
        "scheme" : "bearer",
        "type" : "http"
      }
    }
  },
  "info" : {
    "title" : "OpenAPI definition",
    "version" : "v0"
  },
  "openapi" : "3.1.0",
  "paths" : {
...
    "/api/v2/categories" : {
      "delete" : {
        "operationId" : "v2DeleteAllCategories",
        "responses" : {
          "204" : {
            "description" : "No Content"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "tags" : [ "category-v-2-controller" ]
      },
      "get" : {
        "operationId" : "v2ListCategories",
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/CategoryResponse"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "OK"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "tags" : [ "category-v-2-controller" ]
      },
      "post" : {
        "operationId" : "v2CreateCategory",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CategoryRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CategoryResponse"
                }
              }
            },
            "description" : "Created"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "tags" : [ "category-v-2-controller" ]
      }
    },
    "/api/v2/categories/{id}" : {
      "delete" : {
        "operationId" : "v2DeleteCategory",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "No Content"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "tags" : [ "category-v-2-controller" ]
      },
      "get" : {
        "operationId" : "v2GetCategory",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CategoryResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "tags" : [ "category-v-2-controller" ]
      },
      "put" : {
        "operationId" : "v2UpdateCategory",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CategoryRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CategoryResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "tags" : [ "category-v-2-controller" ]
      }
    },
    "/api/v2/chat/conversations" : {
      "get" : {
        "operationId" : "v2ListConversations",
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/ConversationSummaryDto"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "OK"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "tags" : [ "conversation-v-2-controller" ]
      }
    },
    "/api/v2/chat/conversations/{id}" : {
      "delete" : {
        "operationId" : "v2DeleteConversation",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "No Content"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "tags" : [ "conversation-v-2-controller" ]
      },
      "get" : {
        "operationId" : "v2ConversationMessages",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/ChatMessageDto"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "OK"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "tags" : [ "conversation-v-2-controller" ]
      }
    },
...
    "/budgets" : {
      "delete" : {
        "operationId" : "deleteAll_3",
        "parameters" : [ {
          "in" : "query",
          "name" : "month",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "No Content"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "tags" : [ "budget-controller" ]
      },
      "get" : {
        "operationId" : "list_4",
        "parameters" : [ {
          "in" : "query",
          "name" : "month",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/BudgetResponse"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "OK"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "tags" : [ "budget-controller" ]
      },
      "post" : {
        "operationId" : "create_5",
        "parameters" : [ {
          "in" : "query",
          "name" : "force",
          "required" : false,
          "schema" : {
            "default" : false,
            "type" : "boolean"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/BudgetRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/BudgetResponse"
                }
              }
            },
            "description" : "Created"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "tags" : [ "budget-controller" ]
      }
    },
...
    "/categories" : {
      "delete" : {
        "operationId" : "deleteAll_2",
        "responses" : {
          "204" : {
            "description" : "No Content"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "tags" : [ "category-controller" ]
      },
      "get" : {
        "operationId" : "list_3",
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/CategoryResponse"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "OK"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "tags" : [ "category-controller" ]
      },
      "post" : {
        "operationId" : "create_4",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CategoryRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CategoryResponse"
                }
              }
            },
            "description" : "Created"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "tags" : [ "category-controller" ]
      }
    },
    "/categories/{id}" : {
      "delete" : {
        "operationId" : "delete_3",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "No Content"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "tags" : [ "category-controller" ]
      },
      "get" : {
        "operationId" : "get_3",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CategoryResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "tags" : [ "category-controller" ]
      },
      "put" : {
        "operationId" : "update_4",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CategoryRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CategoryResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "tags" : [ "category-controller" ]
      }
    },
    "/chat/conversations" : {
      "get" : {
        "operationId" : "list_5",
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/ConversationSummaryDto"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "OK"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "tags" : [ "conversation-controller" ]
      }
    },
    "/chat/conversations/{id}" : {
      "delete" : {
        "operationId" : "delete_5",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "No Content"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "tags" : [ "conversation-controller" ]
      },
      "get" : {
        "operationId" : "messages",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/ChatMessageDto"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "OK"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "tags" : [ "conversation-controller" ]
      }
    },
    "/expenses" : {
      "delete" : {
        "operationId" : "deleteAll_1",
        "responses" : {
          "204" : {
            "description" : "No Content"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "tags" : [ "expense-controller" ]
      },
      "get" : {
        "operationId" : "list_2",
        "parameters" : [ {
          "in" : "query",
          "name" : "from",
          "required" : false,
          "schema" : {
            "format" : "date",
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "to",
          "required" : false,
          "schema" : {
            "format" : "date",
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/ExpenseResponse"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "OK"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "tags" : [ "expense-controller" ]
      },
      "post" : {
        "operationId" : "create_3",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ExpenseRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ExpenseResponse"
                }
              }
            },
            "description" : "Created"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "tags" : [ "expense-controller" ]
      }
    },
...
    "/recurring" : {
      "delete" : {
        "operationId" : "deleteAll",
        "responses" : {
          "204" : {
            "description" : "No Content"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "tags" : [ "recurring-expense-controller" ]
      },
      "get" : {
        "operationId" : "findAll",
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/RecurringExpenseResponse"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "OK"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "tags" : [ "recurring-expense-controller" ]
      },
      "post" : {
        "operationId" : "create_1",
        "parameters" : [ {
          "in" : "query",
          "name" : "force",
          "required" : false,
          "schema" : {
            "default" : false,
            "type" : "boolean"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/RecurringExpenseRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RecurringExpenseResponse"
                }
              }
            },
            "description" : "Created"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "tags" : [ "recurring-expense-controller" ]
      }
    },
...
  },
  "servers" : [ {
    "description" : "Generated server url",
    "url" : "http://localhost"
  } ]
}